chore: split class
