+++ src/MagicOnion.OpenTelemetryNew/MagicOnionOpenTelemetryOption.cs
@@ -1,0 +1,22 @@
+﻿using System.Reflection;
+
+namespace MagicOnion.OpenTelemetry
+{
+    public class MagicOnionOpenTelemetryOption
+    {
+        /// <summary>
+        /// Service Name for the app. default is Assembly name.
+        /// </summary>
+        public string ServiceName { get; }
+
+        public MagicOnionOpenTelemetryOption()
+        {
+            ServiceName = Assembly.GetExecutingAssembly().GetName().Name;
+        }
+
+        public MagicOnionOpenTelemetryOption(string serviceName)
+        {
+            ServiceName = serviceName;
+        }
+    }
+}
